requirements.txt
flask==3.0.0
python-docx==1.1.0
lxml==5.1.0